requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.19
langchain_openai==0.2.9
python-dotenv==1.0.1
